Add follow/shake rig to camera
Changed camera setup to use a follow rig and a shake rig to seperate the follow smoothing from any camera effects we might want. Found a smoothing formula thats supposed to be framerate independent.

alpha := 1.0 - exp(-smooth_speed * delta)

also changed it to smooth speed so we dont need the 21 - smoothing. it was confusing.